Finding Error
After getting rid of mallocs and seg faults, the program still seems to be misbehaving. In order to get it work properly, I'm trying to check if there are any logical errors. So, in this commit in recover.c I try to only focus on the positions where jpeg headers are being found, storing them in an array called positions[ ], and printing them. Looks like there is a major issue here, but I can't figure out what. Run the program with card.raw as the first argument.